+++ backend/adapters/services/igdb.py
@@ -13,10 +13,10 @@ from unidecode import unidecode
 
 from adapters.services.igdb_types import Game
 from config import IGDB_CLIENT_ID
-from handler.metadata.base_handler import UniversalPlatformSlug as UPS
 from logger.logger import log
 from utils import get_version
 from utils.context import ctx_aiohttp_session
+from utils.platform_slugs import UniversalPlatformSlug as UPS
 from utils.rate_limiter import RateLimiter
 
 if TYPE_CHECKING:
