+++ backend/alembic/versions/0046_migrate_platform_slugs.py
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 from alembic import op
 
 from config.config_manager import config_manager as cm
-from handler.metadata.base_handler import UniversalPlatformSlug as UPS
+from utils.platform_slugs import UniversalPlatformSlug as UPS
 
 revision = "0046_migrate_platform_slugs"
 down_revision = "0045_roms_metadata_update"
