+++ backend/handler/metadata/hasheous_handler.py
@@ -12,9 +12,9 @@ from logger.logger import log
 from models.rom import RomFile
 from utils import get_version
 from utils.context import ctx_httpx_client
+from utils.platform_slugs import UniversalPlatformSlug as UPS
 
 from .base_handler import BaseRom, MetadataHandler
-from .base_handler import UniversalPlatformSlug as UPS
 from .igdb_handler import (
     IGDB_AGE_RATINGS,
     IGDBMetadata,
