+++ backend/tests/handler/metadata/test_libretro_handler.py
@@ -65,7 +65,7 @@ def test_get_platform_unsupported_platform(handler: LibretroHandler):
 
 def test_platform_list_uses_ups_keys():
     """Every entry in LIBRETRO_PLATFORM_LIST should be a UniversalPlatformSlug."""
-    from handler.metadata.base_handler import UniversalPlatformSlug
+    from utils.platform_slugs import UniversalPlatformSlug
 
     for key in LIBRETRO_PLATFORM_LIST.keys():
         assert isinstance(key, UniversalPlatformSlug)
